refactor(concurrent-bug): extract StubExtensionContext to separate file
